refactor: split quality report generation
