+++ test/quality/report/translations.ts
@@ -1,0 +1,239 @@
+export const REPORT_TRANSLATIONS = {
+	en: {
+		title: "PixelRefiner quality report",
+		groundTruth: "Ground truth",
+		input: "Input",
+		baseline: "Baseline",
+		result: "Result",
+		groundTruthDifference: "Ground-truth difference",
+		baselineDifference: "Baseline difference",
+		backgroundMask: "Background mask",
+		inputKind: "Input kind",
+		route: "Route",
+		confidence: "Confidence (diagnostic)",
+		notAvailable: "not available",
+		warnings: "Warnings",
+		none: "none",
+		topCandidates: "Top candidates",
+		metrics: "Metrics",
+		options: "Options",
+		filterCases: "Filter cases",
+		language: "Language",
+		allStatuses: "All",
+		passed: "passed",
+		failed: "target unmet",
+		preserve: "preserve",
+		refine: "refine",
+		workflow: "workflow",
+		changed: "changed from base branch",
+		improved: "improved from base branch",
+		regressed: "regressed from base branch",
+		unchanged: "unchanged from base branch",
+		new: "new case not in base branch",
+		changedCases: "Cases with differences",
+		allChanges: "All",
+		qualityStatus: "Quality status",
+		changeStatus: "Change status",
+		reportDetails: "Report details",
+		pullRequest: "Pull request",
+		headCommit: "Head",
+		baseCommit: "PR base",
+		baselineCommit: "Baseline snapshot",
+		generatedAt: "Generated",
+		displayConditions: "Showing",
+		casesShown: "cases",
+		changedPixels: "Changed pixels",
+		comparison: "Metric comparison",
+		metric: "Metric",
+		target: "Target",
+		current: "Current",
+		delta: "Delta",
+		verdict: "Verdict",
+		outputSize: "Output size",
+		meanRgbaError: "Mean RGBA error",
+		meanRgbaErrorShort: "Error",
+		processingTime: "Time",
+		exactMatch: "Exact match",
+		exactMatchShort: "Exact",
+		yes: "yes",
+		no: "no",
+		edgeF1: "Edge F1",
+		backgroundMaskIou: "Background mask IoU",
+		smallComponentRetention: "Small component retention",
+		diagnostics: "All images and settings",
+		details: "Details",
+		backToReport: "Back to report",
+		noRegression: "No new quality regression",
+		hasRegression: "Quality regression detected",
+		assertions: {
+			"exact-image-match": "exact image match",
+			"mean-rgba-error": "mean RGBA error",
+			"edge-f1": "edge retention",
+			"background-mask-iou": "background mask",
+			"small-component-retention": "small component retention",
+			"expected-width": "expected width",
+			"expected-height": "expected height",
+			"deterministic-output": "deterministic output",
+			"catastrophic-failure": "catastrophic failure",
+			"output-size": "output size",
+		},
+	},
+	ja: {
+		title: "品質レポート",
+		groundTruth: "期待結果",
+		input: "入力",
+		baseline: "基準結果",
+		result: "処理結果",
+		groundTruthDifference: "期待結果との差分",
+		baselineDifference: "基準結果との差分",
+		backgroundMask: "背景マスク",
+		inputKind: "入力種別",
+		route: "処理ルート",
+		confidence: "信頼度（診断値）",
+		notAvailable: "取得不可",
+		warnings: "警告",
+		none: "なし",
+		topCandidates: "上位候補",
+		metrics: "評価指標",
+		options: "処理設定",
+		filterCases: "ケースを絞り込む",
+		language: "表示言語",
+		allStatuses: "すべて",
+		passed: "合格",
+		failed: "目標未達",
+		preserve: "保持",
+		refine: "復元",
+		workflow: "実行ログ",
+		changed: "base branchから変更",
+		improved: "base branchより改善",
+		regressed: "base branchより悪化",
+		unchanged: "base branchと差分なし",
+		new: "base branchにない新規ケース",
+		changedCases: "差分あり",
+		allChanges: "すべて",
+		qualityStatus: "品質状態",
+		changeStatus: "変化状態",
+		reportDetails: "レポート情報",
+		pullRequest: "プルリクエスト",
+		headCommit: "HEAD",
+		baseCommit: "PRのベース",
+		baselineCommit: "比較基準",
+		generatedAt: "生成日時",
+		displayConditions: "表示条件",
+		casesShown: "件",
+		changedPixels: "変更画素",
+		comparison: "指標の比較",
+		metric: "指標",
+		target: "合格条件",
+		current: "今回",
+		delta: "変化量",
+		verdict: "判定",
+		outputSize: "出力サイズ",
+		meanRgbaError: "RGBA平均誤差",
+		meanRgbaErrorShort: "誤差",
+		processingTime: "時間",
+		exactMatch: "完全一致",
+		exactMatchShort: "一致",
+		yes: "はい",
+		no: "いいえ",
+		edgeF1: "輪郭F1",
+		backgroundMaskIou: "背景マスクIoU",
+		smallComponentRetention: "小要素保持率",
+		diagnostics: "すべての画像と処理設定",
+		details: "詳細",
+		backToReport: "レポートに戻る",
+		noRegression: "新たな品質悪化はありません",
+		hasRegression: "品質の悪化を検出しました",
+		assertions: {
+			"exact-image-match": "画像の完全一致",
+			"mean-rgba-error": "RGBA平均誤差",
+			"edge-f1": "輪郭の保持",
+			"background-mask-iou": "背景マスク",
+			"small-component-retention": "小要素の保持",
+			"expected-width": "期待する幅",
+			"expected-height": "期待する高さ",
+			"deterministic-output": "出力の再現性",
+			"catastrophic-failure": "致命的な失敗",
+			"output-size": "出力サイズ",
+		},
+	},
+	"zh-CN": {
+		title: "PixelRefiner 质量报告",
+		groundTruth: "预期结果",
+		input: "输入",
+		baseline: "基准结果",
+		result: "处理结果",
+		groundTruthDifference: "与预期结果的差异",
+		baselineDifference: "与基准结果的差异",
+		backgroundMask: "背景蒙版",
+		inputKind: "输入类型",
+		route: "处理路径",
+		confidence: "置信度（诊断值）",
+		notAvailable: "不可用",
+		warnings: "警告",
+		none: "无",
+		topCandidates: "候选前三名",
+		metrics: "指标",
+		options: "处理设置",
+		filterCases: "筛选用例",
+		language: "显示语言",
+		allStatuses: "全部",
+		passed: "通过",
+		failed: "未达到目标",
+		preserve: "保留",
+		refine: "优化",
+		workflow: "工作流",
+		changed: "与基础分支不同",
+		improved: "优于基础分支",
+		regressed: "劣于基础分支",
+		unchanged: "与基础分支相同",
+		new: "基础分支中没有的新用例",
+		changedCases: "有差异的用例",
+		allChanges: "全部",
+		qualityStatus: "质量状态",
+		changeStatus: "变更状态",
+		reportDetails: "报告信息",
+		pullRequest: "拉取请求",
+		headCommit: "HEAD",
+		baseCommit: "PR 基础提交",
+		baselineCommit: "比较基准",
+		generatedAt: "生成时间",
+		displayConditions: "显示条件",
+		casesShown: "个用例",
+		changedPixels: "变更像素",
+		comparison: "指标比较",
+		metric: "指标",
+		target: "目标",
+		current: "当前",
+		delta: "变化量",
+		verdict: "判定",
+		outputSize: "输出尺寸",
+		meanRgbaError: "RGBA 平均误差",
+		meanRgbaErrorShort: "误差",
+		processingTime: "时间",
+		exactMatch: "完全匹配",
+		exactMatchShort: "匹配",
+		yes: "是",
+		no: "否",
+		edgeF1: "边缘 F1",
+		backgroundMaskIou: "背景蒙版 IoU",
+		smallComponentRetention: "小组件保留率",
+		diagnostics: "所有图像和处理设置",
+		details: "详情",
+		backToReport: "返回报告",
+		noRegression: "未发现新的质量下降",
+		hasRegression: "检测到质量下降",
+		assertions: {
+			"exact-image-match": "图像完全匹配",
+			"mean-rgba-error": "RGBA 平均误差",
+			"edge-f1": "边缘保留",
+			"background-mask-iou": "背景蒙版",
+			"small-component-retention": "小组件保留",
+			"expected-width": "预期宽度",
+			"expected-height": "预期高度",
+			"deterministic-output": "输出可重复性",
+			"catastrophic-failure": "灾难性失败",
+			"output-size": "输出尺寸",
+		},
+	},
+} as const;
